refactor(facade): move isPromise to core private
