+++ modules/@angular/core/src/application_init.ts
@@ -6,7 +6,7 @@
  * found in the LICENSE file at https://angular.io/license
  */
 
-import {isPromise} from '../src/facade/lang';
+import {isPromise} from '../src/util/lang';
 
 import {Inject, Injectable, OpaqueToken, Optional} from './di';
