+++ modules/@angular/core/testing/testing_internal.ts
@@ -8,7 +8,8 @@
 
 import {AsyncTestCompleter} from './async_test_completer';
 import {StringMapWrapper} from './facade/collection';
-import {Math, global, isPromise} from './facade/lang';
+import {Math, global} from './facade/lang';
+import {isPromise} from './private_import_core';
 import {TestBed, getTestBed, inject} from './test_bed';
 
 export {AsyncTestCompleter} from './async_test_completer';
