+++ modules/@angular/facade/src/lang.ts
@@ -113,12 +113,6 @@ export function isStrictStringMap(obj: any): boolean {
   return isStringMap(obj) && Object.getPrototypeOf(obj) === STRING_MAP_PROTO;
 }
 
-export function isPromise(obj: any): boolean {
-  // allow any Promise/A+ compliant thenable.
-  // It's up to the caller to ensure that obj.then conforms to the spec
-  return isPresent(obj) && isFunction(obj.then);
-}
-
 export function isArray(obj: any): boolean {
   return Array.isArray(obj);
 }
