+++ modules/@angular/facade/test/lang_spec.ts
@@ -6,7 +6,7 @@
  * found in the LICENSE file at https://angular.io/license
  */
 
-import {NumberWrapper, StringWrapper, escapeRegExp, hasConstructor, isPresent, isPromise, resolveEnumToken} from '../src/lang';
+import {NumberWrapper, StringWrapper, escapeRegExp, hasConstructor, isPresent, resolveEnumToken} from '../src/lang';
 
 enum UsefulEnum {
   MyToken,
@@ -151,24 +151,6 @@ export function main() {
 
       it('should be false for subtypes',
          () => { expect(hasConstructor(new MySubclass(), MySuperclass)).toEqual(false); });
-    });
-  });
-  describe('isPromise', () => {
-    it('should be true for native Promises',
-       () => expect(isPromise(Promise.resolve(true))).toEqual(true));
-
-    it('should be true for thenables',
-       () => expect(isPromise({then: function() {}})).toEqual(true));
-
-    it('should be false if "then" is not a function',
-       () => expect(isPromise({then: 0})).toEqual(false));
-
-    it('should be false if the argument has no "then" function',
-       () => expect(isPromise({})).toEqual(false));
-
-    it('should be false if the argument is undefined or null', () => {
-      expect(isPromise(undefined)).toEqual(false);
-      expect(isPromise(null)).toEqual(false);
     });
   });
 }
