+++ modules/@angular/forms/src/model.ts
@@ -12,7 +12,8 @@ import {composeAsyncValidators, composeValidators} from './directives/shared';
 import {AsyncValidatorFn, ValidatorFn} from './directives/validators';
 import {EventEmitter, Observable} from './facade/async';
 import {ListWrapper, StringMapWrapper} from './facade/collection';
-import {isBlank, isPresent, isPromise, isStringMap, normalizeBool} from './facade/lang';
+import {isBlank, isPresent, isStringMap, normalizeBool} from './facade/lang';
+import {isPromise} from './private_import_core';
